backend/requirements.txt
transformers
torch
flask
pandas
openpyxl
psycopg2-binary
python-dotenv


